Install certs and make windows with cmake
Arm build fails due to certificate errors. Either we don't care about
the certs (bad) or we try to install the correct certs (good) so it
doesn't fail on that.

Visual Studio is not like the others. You apparently don't get a
makefile when using VS to build things. So instead of using `make` you
should use `cmake --build .` to actually build it. Let's give that a go.